+++ src/main/java/codesquad/server/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.server.http;
 
 public enum HttpMethod {
     HEAD, GET, POST, PUT, DELETE, OPTIONS, TRACE
