+++ src/main/java/codesquad/server/http/HttpRequest.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.server.http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
