+++ src/main/java/codesquad/server/http/Params.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.server.http;
 
 import codesquad.exception.HttpException;
