+++ src/main/java/codesquad/server/router/StaticFileProcessor.java
@@ -1,9 +1,9 @@
-package codesquad.server;
+package codesquad.server.router;
 
 import codesquad.exception.HttpException;
-import codesquad.http.ContentType;
-import codesquad.http.HttpResponse;
-import codesquad.http.HttpStatus;
+import codesquad.server.http.ContentType;
+import codesquad.server.http.HttpResponse;
+import codesquad.server.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
