+++ src/test/java/codesquad/server/http/StaticFileProcessorTest.java
@@ -1,7 +1,9 @@
-package codesquad.http;
+package codesquad.server.http;
 
-import codesquad.server.StaticFilePathManager;
-import codesquad.server.StaticFileProcessor;
+import codesquad.server.router.StaticFilePathManager;
+import codesquad.server.router.StaticFileProcessor;
+import codesquad.server.http.ContentType;
+import codesquad.server.http.HttpResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
