+++ src/test/java/codesquad/usecase/UserUsecaseTest.java
@@ -1,8 +1,8 @@
 package codesquad.usecase;
 
-import codesquad.http.HttpMethod;
-import codesquad.http.HttpRequest;
-import codesquad.http.HttpResponse;
+import codesquad.server.http.HttpMethod;
+import codesquad.server.http.HttpRequest;
+import codesquad.server.http.HttpResponse;
 import codesquad.storage.UserStorage;
 import codesquad.util.HttpRequestUtil;
 import org.assertj.core.api.Assertions;
